phase-1/short-url/application/auth/app_files/requirements.txt
Flask
Flask-MySQLdb
mysql-connector-python
PyMySQL
cryptography
prometheus_client
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-pymysql